Fix failing device tests
