+++ src/sync_stream.rs
@@ -889,7 +889,6 @@ pub use tokio_stream::EventStream;
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tests::{get_test_device, key_event};
 
     fn result_events_iter(
         events: &[input_event],
@@ -956,24 +955,5 @@ mod tests {
         assert_eq!(next(), (Err(false), None));
         assert_eq!(next(), (Err(false), None));
         assert_eq!(next(), (Err(false), None));
-    }
-
-    #[test]
-    pub fn test_get_key_state() -> Result<(), Box<dyn std::error::Error>> {
-        let (input, mut output) = get_test_device()?;
-
-        output.emit(&[key_event(KeyCode::KEY_DOT, 1)])?;
-
-        assert_eq!(1, input.get_key_state()?.iter().count());
-        assert!(input
-            .get_key_state()?
-            .iter()
-            .all(|e| e.code() == KeyCode::KEY_DOT.code()));
-
-        output.emit(&[key_event(KeyCode::KEY_DOT, 0)])?;
-
-        assert_eq!(0, input.get_key_state()?.iter().count());
-
-        Ok(())
     }
 }
